+++ scm-ui/ui-components/src/MarkdownView.stories.tsx
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
  * SOFTWARE.
  */
-import React from "react";
+import React, { FC } from "react";
 import { storiesOf } from "@storybook/react";
 import MarkdownView from "./MarkdownView";
 import styled from "styled-components";
@@ -29,20 +29,22 @@ import styled from "styled-components";
 import TestPage from "./__resources__/test-page.md";
 import MarkdownWithoutLang from "./__resources__/markdown-without-lang.md";
 import MarkdownXmlCodeBlock from "./__resources__/markdown-xml-codeblock.md";
+import MarkdownUmlCodeBlock from "./__resources__/markdown-uml-codeblock.md";
 import MarkdownInlineXml from "./__resources__/markdown-inline-xml.md";
 import MarkdownLinks from "./__resources__/markdown-links.md";
 import MarkdownCommitLinks from "./__resources__/markdown-commit-link.md";
 import Title from "./layout/Title";
 import { Subtitle } from "./layout";
 import { MemoryRouter } from "react-router-dom";
+import { Binder, BinderContext } from "@scm-manager/ui-extensions";
 
 const Spacing = styled.div`
   padding: 2em;
 `;
 
 storiesOf("MarkdownView", module)
-  .addDecorator(story => <MemoryRouter initialEntries={["/"]}>{story()}</MemoryRouter>)
-  .addDecorator(story => <Spacing>{story()}</Spacing>)
+  .addDecorator((story) => <MemoryRouter initialEntries={["/"]}>{story()}</MemoryRouter>)
+  .addDecorator((story) => <Spacing>{story()}</Spacing>)
   .add("Default", () => <MarkdownView content={TestPage} skipHtml={false} />)
   .add("Code without Lang", () => <MarkdownView content={MarkdownWithoutLang} skipHtml={false} />)
   .add("Xml Code Block", () => <MarkdownView content={MarkdownXmlCodeBlock} />)
@@ -54,4 +56,23 @@ storiesOf("MarkdownView", module)
     </>
   ))
   .add("Links", () => <MarkdownView content={MarkdownLinks} basePath="/" />)
-  .add("Commit Links", () => <MarkdownView content={MarkdownCommitLinks} />);
+  .add("Commit Links", () => <MarkdownView content={MarkdownCommitLinks} />)
+  .add("Custom code renderer", () => {
+    const binder = new Binder("custom code renderer");
+    const Container: FC<{ value: string }> = ({ value }) => {
+      return (
+        <div>
+          <h4 style={{ border: "1px dashed lightgray", padding: "2px" }}>
+            To render plantuml as images within markdown, please install the scm-markdown-plantuml-plguin
+          </h4>
+          <pre>{value}</pre>
+        </div>
+      );
+    };
+    binder.bind("markdown-renderer.code.uml", Container);
+    return (
+      <BinderContext.Provider value={binder}>
+        <MarkdownView content={MarkdownUmlCodeBlock} />
+      </BinderContext.Provider>
+    );
+  });